fdk: Fix non-centered source position
When the source position was not located exactly on the line through
the detector center parallel to the detector normal, the pre-weighting
could be slightly off.

This has now been fixed.

Also, some changes could be made to the pre-weighting that did not
result in test failure. This possibility has been further minimized.