Remove smoothing in TV functional
In d9741ef, the TV functional was changed to use a
smoothing parameter to avoid possible division by
zero if gradients are calculated. However, the
smoothing makes the Chambolle-Pock algorithm not
strictly applicable. Instead of smoothing, this
commit addresses the division problem by replacing
the gradient at zero with its subgradient (zero
tensor in this case). This is also the default
behavior of `torch.norm` function.